go/g215/215.数组中的第k个最大元素_test.go
/*
 * @lc app=leetcode.cn id=215 lang=golang
 *
 * [215] 数组中的第K个最大元素
 *
 * https://leetcode.cn/problems/kth-largest-element-in-an-array/description/
 *
 * algorithms
 * Medium (64.70%)
 * Likes:    1671
 * Dislikes: 0
 * Total Accepted:    627.3K
 * Total Submissions: 969.4K
 * Testcase Example:  '[3,2,1,5,6,4]\n2'
 *
 * 给定整数数组 nums 和整数 k，请返回数组中第 k 个最大的元素。
 *
 * 请注意，你需要找的是数组排序后的第 k 个最大的元素，而不是第 k 个不同的元素。
 *
 *
 *
 * 示例 1:
 *
 *
 * 输入: [3,2,1,5,6,4] 和 k = 2
 * 输出: 5
 *
 *
 * 示例 2:
 *
 *
 * 输入: [3,2,3,1,2,4,5,5,6] 和 k = 4
 * 输出: 4
 *
 *
 *
 * 提示：
 *
 *
 * 1
 * -10^4
 *
 *
 */

package g215

import "testing"

func Test_findKthLargestV4(t *testing.T) {
	type args struct {
		nums []int
		k    int
	}
	tests := []struct {
		name string
		args args
		want int
	}{
		// {
		// 	name: "test1",
		// 	args: args{
		// 		nums: []int{3, 2, 3, 1, 2, 4, 5, 5, 6},
		// 		k:    4,
		// 	},
		// },
		// {
		// 	name: "test2",
		// 	args: args{
		// 		nums: []int{32, 23, 9993, 1, 244, 476, 8885, 5, 666666},
		// 		k:    4,
		// 	},
		// },
		{
			name: "test3",
			args: args{
				nums: []int{-1, -100, 2, 0, 80, -9999, 777777},
				k:    2,
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := findKthLargestV10(tt.args.nums, tt.args.k); got != tt.want {
				t.Errorf("findKthLargest() = %v, want %v", got, tt.want)
			}
		})
	}
}
